Setting up WebPack
Lecture 50
