+++ indecision-app/public/index.html
@@ -6,9 +6,6 @@
 </head>
 <body>
     <div id="app"></div>
-    <script src="https://unpkg.com/react@16.0.0/umd/react.development.js"></script>
-    <script src="https://unpkg.com/react-dom@16.0.0/umd/react-dom.development.js"></script>
-    <!--<script src="/scripts/build-it-visible.js"></script>-->
-    <script src="/scripts/app.js"></script>
+    <script src="/bundle.js"></script>
 </body>
 </html>
