+++ indecision-app/webpack.config.js
@@ -1,0 +1,10 @@
+const path = require('path');
+
+// Exposes object to other file
+module.exports = {
+    entry: './src/app.js',
+    output: {
+        path: path.join(__dirname, 'public'),
+        filename: 'bundle.js'
+    }
+};
